Implement feature to allow hidden files in local copy
