+++ config/config_as_cty.go
@@ -127,10 +127,11 @@ func TerragruntConfigAsCty(config *TerragruntConfig) (cty.Value, error) {
 // ctyTerraformConfig is an alternate representation of TerraformConfig that converts internal blocks into a map that
 // maps the name to the underlying struct, as opposed to a list representation.
 type ctyTerraformConfig struct {
-	ExtraArgs   map[string]TerraformExtraArguments `cty:"extra_arguments"`
-	Source      *string                            `cty:"source"`
-	BeforeHooks map[string]Hook                    `cty:"before_hook"`
-	AfterHooks  map[string]Hook                    `cty:"after_hook"`
+	ExtraArgs     map[string]TerraformExtraArguments `cty:"extra_arguments"`
+	Source        *string                            `cty:"source"`
+	IncludeInCopy *[]string                          `cty:"include_in_copy"`
+	BeforeHooks   map[string]Hook                    `cty:"before_hook"`
+	AfterHooks    map[string]Hook                    `cty:"after_hook"`
 }
 
 // Serialize TerraformConfig to a cty Value, but with maps instead of lists for the blocks.
@@ -140,10 +141,11 @@ func terraformConfigAsCty(config *TerraformConfig) (cty.Value, error) {
 	}
 
 	configCty := ctyTerraformConfig{
-		Source:      config.Source,
-		ExtraArgs:   map[string]TerraformExtraArguments{},
-		BeforeHooks: map[string]Hook{},
-		AfterHooks:  map[string]Hook{},
+		Source:        config.Source,
+		IncludeInCopy: config.IncludeInCopy,
+		ExtraArgs:     map[string]TerraformExtraArguments{},
+		BeforeHooks:   map[string]Hook{},
+		AfterHooks:    map[string]Hook{},
 	}
 
 	for _, arg := range config.ExtraArgs {
