+++ test/fixture-download/local-with-allowed-hidden/live/terragrunt.hcl
@@ -1,0 +1,4 @@
+terraform {
+  source          = "../modules"
+  include_in_copy = [".nonce"]
+}
