+++ test/fixture-download/local-with-allowed-hidden/modules/.nonce
@@ -1,0 +1,1 @@
+Hello world
