+++ test/fixture-download/local-with-allowed-hidden/modules/main.tf
@@ -1,0 +1,3 @@
+output "text" {
+  value = trimspace(file(".nonce"))
+}
